+++ web/modules/custom/chat/src/Controller/ChatStreamController.php
@@ -2,11 +2,13 @@
 
 namespace Drupal\chat\Controller;
 
+use Drupal\chat\Model\ChatCallbackData;
 use Drupal\Core\Controller\ControllerBase;
 use Drupal\llm_services\Model\Message;
 use Drupal\llm_services\Model\MessageRoles;
 use Drupal\llm_services\Model\Payload;
 use Drupal\llm_services\Plugin\LLModelProviderManager;
+use GuzzleHttp\Exception\InvalidArgumentException;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -47,14 +49,15 @@ class ChatStreamController extends ControllerBase {
    *   Stream with text content from the LLM.
    *
    * @throws \JsonException
+   * @throws \Drupal\Component\Plugin\Exception\PluginException
    */
   public function callback(Request $request): StreamedResponse {
-    // @todo Validate data in request.
-    $data = json_decode($request->getContent(), associative: TRUE, flags: JSON_THROW_ON_ERROR);
+    $json = json_decode($request->getContent(), associative: TRUE, flags: JSON_THROW_ON_ERROR);
+    $data = $this->mapJsonToData($json);
 
-    $provider = $this->providerManager->createInstance($data['provider']);
+    $provider = $this->providerManager->createInstance($data->provider);
 
-    $expire = $data['context_expire'];
+    $expire = $data->contextExpire;
     $cid = \Drupal::service('session')->getId();
     $cached = \Drupal::cache('chat')->get($cid);
     if ($cached) {
@@ -62,18 +65,24 @@ class ChatStreamController extends ControllerBase {
     }
     else {
       $payload = new Payload();
-      $payload->setModel($data['model'])
-        ->addOption('temperature', $data['temperature'])
-        ->addOption('top_k', $data['top_k'])
-        ->addOption('top_p', $data['top_p']);
+      $payload->setModel($data->model)
+        ->addOption('temperature', $data->temperature)
+        ->addOption('top_k', $data->topK)
+        ->addOption('top_p', $data->topP);
       $msg = new Message();
       $msg->role = MessageRoles::System;
-      $msg->content = $data['system_prompt'];
+      $msg->content = $data->systemPrompt;
       $payload->addMessage($msg);
     }
+
+    // Enforce context length, number of message in the payload based on
+    // configuration.
+    $this->enforceContextLength($payload, (int) $data->contextLength);
+
+    // Add the newest question.
     $msg = new Message();
     $msg->role = MessageRoles::User;
-    $msg->content = $data['prompt'];
+    $msg->content = $data->prompt;
     $payload->addMessage($msg);
 
     return new StreamedResponse(
@@ -108,7 +117,10 @@ class ChatStreamController extends ControllerBase {
         // text, even though we send clean text. If changed stream stops
         // working.
         'Content-Type' => 'application/json',
+        // Ensure nginx and proxy do not cache.
         'X-Accel-Buffering' => 'no',
+        // Ensure browser do not cache.
+        'Cache-Control' => 'no-cache, no-store, private',
       ]
     );
   }
@@ -127,6 +139,93 @@ class ChatStreamController extends ControllerBase {
     \Drupal::cache('chat')->delete($cid);
 
     return new Response('', Response::HTTP_NO_CONTENT);
+  }
+
+  /**
+   * Enforce the context length of a payload's messages.
+   *
+   * @param \Drupal\llm_services\Model\Payload $payload
+   *   The payload to enforce the context length of.
+   * @param int $context_length
+   *   The desired context length. Default: 10.
+   */
+  private function enforceContextLength(Payload $payload, int $context_length = 10): void {
+    $messages = $payload->getMessages();
+
+    // Length times two (every question has an answerer) + system prompt.
+    $max = ($context_length * 2) + 1;
+
+    // If count doesn't exceed max, no need to enforce context length.
+    if (count($messages) <= $max) {
+      return;
+    }
+
+    // Remove the system message.
+    $system = array_shift($messages);
+
+    // Remove messages exceeding the max limit. From the start of the array.
+    $messages = array_slice(
+      array: array_reverse($messages),
+      offset: 0,
+      length: $max - 1
+    );
+
+    // Reverse back to correct order and add the system message as the first
+    // message.
+    $messages = array_reverse($messages);
+    array_unshift($messages, $system);
+
+    // Override messages in the payload.
+    $payload->setMessages($messages);
+  }
+
+  /**
+   * Create a callback payload from JSON.
+   *
+   * @param array $json
+   *   The JSON data to create the callback payload from.
+   *
+   * @return \Drupal\Chat\Model\ChatCallbackData
+   *   The created callback payload object.
+   *
+   * @throws \InvalidArgumentException
+   *   If the JSON given is not valid.
+   */
+  private function mapJsonToData(array $json): ChatCallbackData {
+    $keys = [
+      'provider',
+      'model',
+      'prompt',
+      'system_prompt',
+      'temperature',
+      'top_k',
+      'top_p',
+      'context_expire',
+      'context_length',
+    ];
+
+    // All keys exist.
+    $common = array_intersect_key(array_flip($keys), $json);
+    if (count($keys) !== count($common)) {
+      throw new InvalidArgumentException('Request data is not valid');
+    }
+
+    try {
+      return new ChatCallbackData(
+        provider: $json['provider'],
+        model: $json['model'],
+        prompt: $json['prompt'],
+        systemPrompt: $json['system_prompt'],
+        temperature: (float) $json['temperature'],
+        topK: (int) $json['top_k'],
+        topP: (float) $json['top_p'],
+        contextExpire: (int) $json['context_expire'],
+        contextLength: (int) $json['context_length'],
+      );
+    }
+    catch (\TypeError $exception) {
+      throw new InvalidArgumentException('Request data is not valid: ' . $exception->getMessage(), $exception->getCode(), $exception);
+    }
   }
 
 }