+++ web/modules/custom/chat/src/Plugin/Block/ChatBlock.php
@@ -7,6 +7,7 @@ use Drupal\Core\Block\BlockBase;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
 use Drupal\Core\StringTranslation\TranslatableMarkup;
+use Drupal\Core\Url;
 use Drupal\llm_services\Plugin\LLModelProviderManager;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -57,18 +58,33 @@ class ChatBlock extends BlockBase implements ContainerFactoryPluginInterface {
    * {@inheritdoc}
    */
   public function build(): array {
+    $streamUrl = Url::fromRoute('chat.stream');
+    $resetUrl = Url::fromRoute('chat.reset');
     return [
-      '#theme' => 'chat_block',
-      '#provider_name' => $this->configuration['provider_name'],
-      '#models' => $this->configuration['models'],
-      '#system_prompt' => $this->configuration['system_prompt'],
-      '#temperature' => $this->configuration['temperature'],
-      '#top_k' => $this->configuration['top_k'],
-      '#top_p' => $this->configuration['top_p'],
-      '#context_expire' => $this->configuration['context_expire'],
+      '#theme' => 'chat',
+      '#ui' => [
+        'id' => $this->configuration['ui']['id'],
+        'buttons' => $this->configuration['ui']['buttons'],
+        'models' => $this->configuration['models'],
+      ],
       '#attached' => [
         'library' => [
           'chat/chat',
+        ],
+        'drupalSettings' => [
+          'chat' => [
+            'id' => $this->configuration['ui']['id'],
+            'callback_path' => $streamUrl->toString(),
+            'reset_path' => $resetUrl->toString(),
+            'provider_name' => $this->configuration['provider_name'],
+            'system_prompt' => $this->configuration['system_prompt'],
+            'temperature' => $this->configuration['temperature'],
+            'top_k' => $this->configuration['top_k'],
+            'top_p' => $this->configuration['top_p'],
+            'context_expire' => $this->configuration['context_expire'],
+            'context_length' => $this->configuration['context_length'],
+            'waiter_svg' => \Drupal::service('extension.list.module')->getPath('chat') . '/svg/wait.svg',
+          ],
         ],
       ],
     ];
@@ -86,6 +102,11 @@ class ChatBlock extends BlockBase implements ContainerFactoryPluginInterface {
       'top_k' => 40,
       'top_p' => 0.9,
       'context_expire' => 3600,
+      'context_length' => 10,
+      'ui' => [
+        'id' => 'jsChat',
+        'buttons' => FALSE,
+      ],
     ];
   }
 
@@ -144,6 +165,26 @@ class ChatBlock extends BlockBase implements ContainerFactoryPluginInterface {
       '#required' => TRUE,
     ];
 
+    $form['ui'] = [
+      '#type' => 'details',
+      '#title' => $this->t('User interface tweaks'),
+      '#open' => TRUE,
+    ];
+
+    $form['ui']['id'] = [
+      '#type' => 'textfield',
+      '#title' => $this->t('Chat ID'),
+      '#description' => $this->t('If inserting more that one chat block. Set unique ID for the chat window here.'),
+      '#default_value' => $this->configuration['ui']['id'],
+    ];
+
+    $form['ui']['buttons'] = [
+      '#type' => 'checkbox',
+      '#title' => $this->t('Enable minimize/close buttons'),
+      '#description' => $this->t('If checked, minimize buttons will be enabled.'),
+      '#default_value' => $this->configuration['ui']['buttons'],
+    ];
+
     $form['tune'] = [
       '#type' => 'details',
       '#title' => $this->t('Fine-tune the chat'),
@@ -183,6 +224,15 @@ class ChatBlock extends BlockBase implements ContainerFactoryPluginInterface {
       '#min' => 0,
     ];
 
+    $form['tune']['context_length'] = [
+      '#type' => 'number',
+      '#title' => $this->t('Context length'),
+      '#description' => $this->t('The number of user message to send to the LLM as context.'),
+      '#default_value' => $this->configuration['context_length'],
+      '#min' => 0,
+      '#max' => 25,
+    ];
+
     return $form;
   }
 
@@ -213,6 +263,9 @@ class ChatBlock extends BlockBase implements ContainerFactoryPluginInterface {
     $this->configuration['top_k'] = $values['tune']['top_k'];
     $this->configuration['top_p'] = $values['tune']['top_p'];
     $this->configuration['context_expire'] = $values['tune']['context_expire'];
+    $this->configuration['context_length'] = $values['tune']['context_length'];
+    $this->configuration['ui']['buttons'] = $values['ui']['buttons'];
+    $this->configuration['ui']['id'] = $values['ui']['id'];
   }
 
 }